Lessons/Lesson 15/requirements.txt
gradio==3.50.2
pandas
numpy
Pillow
pytesseract
PyPDF2
python-dotenv 